+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/global/infrastructure/feign/client/GithubOAuthApiClient.kt
@@ -1,6 +1,6 @@
-package entry.dsm.gitauth.equusgithubauth.global.external.github.presentation.controller
+package entry.dsm.gitauth.equusgithubauth.global.infrastructure.feign.client
 
-import entry.dsm.gitauth.equusgithubauth.global.external.github.presentation.dto.response.LoginAccessTokenResponse
+import entry.dsm.gitauth.equusgithubauth.global.infrastructure.feign.client.dto.response.LoginAccessTokenResponse
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestParam
