+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/global/infrastructure/feign/client/dto/response/GithubOrganizationResponse.kt
@@ -1,4 +1,4 @@
-package entry.dsm.gitauth.equusgithubauth.global.external.github.presentation.dto.response
+package entry.dsm.gitauth.equusgithubauth.global.infrastructure.feign.client.dto.response
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
